Commands.hpp file added
